refactor: Organise cards code frontend
