+++ frontend/src/components/StubCards.jsx
